move peakgroup to search
