+++ alphadia/search/peakgroup/config_df.py
@@ -12,8 +12,7 @@ logger = logging.getLogger()
 # TODO rename Config -> "JITConfig" to avoid confusion with the Config class in alphadia.utilities.config ?
 @nb.experimental.jitclass()
 class HybridCandidateConfigJIT:
-    """
-    Numba compatible config object for the HybridCandidate class.
+    """Numba compatible config object for the HybridCandidate class.
     Please see the documentation of the HybridCandidateConfig class for more information on the parameters and their default values.
     """
