+++ ruff-lint-strict.toml
@@ -30,7 +30,7 @@ exclude = [
     "alphadia/_fdrx/plotting.py",
     "alphadia/_fdrx/base.py",
     "alphadia/numba/*.py",
-    "alphadia/peakgroup/*.py",
+    "alphadia/search/**.py",
     "alphadia/plotting/*.py",
     "alphadia/transferlearning/*.py",
     "alphadia/utilities/*.py",
